Flytter kuttePerioderVedOpphør til BeregningsresultatLæremidler for å gjøre den testbar
